requirements.txt
Flask==3.0.3
python-dotenv==1.0.1
PyPDF2==3.0.1
Pillow==10.4.0
PyMuPDF==1.24.9
pytesseract==0.3.13
torch==2.6.0
transformers==4.57.1
requests